farmxpert/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.9.2
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
structlog==24.1.0
redis[hiredis]==5.0.3
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
orjson==3.10.7
PyYAML==6.0.2
google-generativeai==0.3.2
asyncio==3.4.3
aiohttp==3.12.1
python-multipart==0.0.6
alembic==1.13.1
PyJWT==2.8.0
email-validator==2.1.0

packaging==24.2
requests==2.32.3
numpy==1.26.4
pandas==2.2.3
catboost==1.2.7
